redux - commit definitely in wrong branch - convert to MVC
