+++ server/config/consts.js
@@ -1,0 +1,4 @@
+import path from 'path';
+
+export const distPath = path.join(__dirname, '..', '..', 'dist');
+export const indexPath = path.join(__dirname, '..', '..', 'dist', 'index.html');
