+++ server/controllers/callback.js
@@ -1,0 +1,64 @@
+import querystring from 'querystring';
+import {client_id, client_secret, redirect_uri} from '../config/auth';
+import request from 'request';
+
+
+export const handleCallback = (req, res) => {
+  let stateKey = 'spotify_auth_state';
+  let code = req.query.code || null;
+  let state = req.query.state || null;
+
+  let storedState = req.cookies ? req.cookies[stateKey] : null;
+
+  if (state === null || state !== storedState) {
+    res.redirect('/#' +
+      querystring.stringify({
+        error: 'state_mismatch'
+      }));
+  } else {
+    res.clearCookie(stateKey);
+    let authOptions = {
+      url: 'https://accounts.spotify.com/api/token',
+      form: {
+        code: code,
+        redirect_uri: redirect_uri,
+        grant_type: 'authorization_code'
+      },
+      headers: {
+        'Authorization': 'Basic ' + (new Buffer(client_id + ':' + client_secret).toString('base64'))
+      },
+      json: true
+    };
+
+    request.post(authOptions, (error, response, body) => {
+      if (!error && response.statusCode === 200) {
+
+        let access_token = body.access_token,
+          refresh_token = body.refresh_token;
+
+        let options = {
+          url: 'https://api.spotify.com/v1/me',
+          headers: {'Authorization': 'Bearer ' + access_token},
+          json: true
+        };
+
+        // use the access token to access the Spotify Web API
+        request.get(options, (error, response, body) => {
+          console.log(body);
+        });
+
+        // we can also pass the token to the browser to make requests from there
+        res.redirect('/loginSuccess/#' +
+          querystring.stringify({
+            access_token: access_token,
+            refresh_token: refresh_token
+          }));
+      } else {
+        res.redirect('/#' +
+          querystring.stringify({
+            error: 'invalid_token'
+          }));
+      }
+    });
+  }
+};
