+++ server/controllers/refreshToken.js
@@ -1,0 +1,25 @@
+import {client_id, client_secret} from '../config/auth';
+import request from 'request';
+
+export const refreshHandler = (req, res) => {
+  // requesting access token from refresh token
+  const refresh_token = req.query.refresh_token;
+  let authOptions = {
+    url: 'https://accounts.spotify.com/api/token',
+    headers: {'Authorization': 'Basic ' + (new Buffer(client_id + ':' + client_secret).toString('base64'))},
+    form: {
+      grant_type: 'refresh_token',
+      refresh_token: refresh_token
+    },
+    json: true
+  };
+
+  request.post(authOptions, (error, response, body) => {
+    if (!error && response.statusCode === 200) {
+      const access_token = body.access_token;
+      res.send({
+        'access_token': access_token
+      });
+    }
+  });
+};
