+++ server/routes/callback.js
@@ -1,0 +1,8 @@
+import express from 'express';
+import {handleCallback} from '../controllers/callback';
+
+const router = express.Router();
+
+router.get('/callback', handleCallback);
+
+export default router;
