+++ server/routes/login.js
@@ -1,0 +1,8 @@
+import express from 'express';
+import {handleLogin} from '../controllers/login';
+
+const router = express.Router();
+
+router.get('/login', handleLogin);
+
+export default router;
