+++ server/routes/root.js
@@ -1,0 +1,9 @@
+import express from 'express';
+import {indexPath} from '../config/consts';
+const router = express.Router();
+
+router.get('/*', (req, res) => {
+  res.sendFile(indexPath)
+});
+export default router;
+
